no zap XS test dirs if true env EUMM_KEEP_TESTDIRS

+++ t/lib/MakeMaker/Test/Setup/XS.pm
@@ -456,7 +456,11 @@ sub run_tests {
   }
 
   chdir File::Spec->updir or die;
-  ok rmtree($dir), "teardown $dir";
+  if ($ENV{EUMM_KEEP_TESTDIRS}) {
+    ok 1, "don't teardown $dir";
+  } else {
+    ok rmtree($dir), "teardown $dir";
+  }
 }
 
 1;
